# Conflicts:
#	eperusteet/eperusteet-service/src/main/java/db/migration/V0_104__oppiaineen_tunniste.java
#	eperusteet/eperusteet-service/src/main/java/db/migration/V0_161__tutkinnon_rakenne_tunniste_paivitys.java
#	eperusteet/eperusteet-service/src/main/java/db/migration/V0_20230503101700__julkaisu_dokumentit_fix.java
#	eperusteet/eperusteet-service/src/main/java/db/migration/V0_98__tunniste.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/AbstractRakenneOsaDeserializer.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/AsyncConfig.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/DokumenttiAsyncConfig.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/EnumToUpperCaseEditor.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/ExceptionHandlingConfig.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/InitBinderControllerAdvice.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/InternalApi.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/MappingModule.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/MvcConfiguration.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/PerusteenOsaUpdateDtoDeserializer.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/ReferenceNamingStrategy.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/config/SwaggerConfig.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/domain/Osaamistaso.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/resource/TiedoteController.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/resource/peruste/PerusopetusOppiaineLockController.java
#	eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/resource/util/ProfilingInterceptor.java
#	eperusteet/eperusteet-service/src/test/java/fi/vm/sade/eperusteet/service/ArviointiServiceIT.java
#	eperusteet/eperusteet-service/src/test/java/fi/vm/sade/eperusteet/service/TutkinnonOsaJulkisetIT.java

+++ eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/repository/custom/AmmattitaitovaatimusRepositoryImpl.java
@@ -29,7 +29,7 @@ public class AmmattitaitovaatimusRepositoryImpl implements AmmattitaitovaatimusR
         TypedQuery<TutkinnonOsaViite> query = getTosaQuery(pquery);
         TypedQuery<Long> tosaCountQuery = getTosaCountQuery(pquery);
         if (page != null) {
-            query.setFirstResult(page.getOffset());
+            query.setFirstResult(Long.valueOf(page.getOffset()).intValue());
             query.setMaxResults(page.getPageSize());
         }
         Page<TutkinnonOsaViite> result = new PageImpl<>(
@@ -82,7 +82,7 @@ public class AmmattitaitovaatimusRepositoryImpl implements AmmattitaitovaatimusR
         TypedQuery<Long> countQuery = getCountQuery(pquery);
         TypedQuery<Perusteprojekti> query = getQuery(pquery);
         if (page != null) {
-            query.setFirstResult(page.getOffset());
+            query.setFirstResult(Long.valueOf(page.getOffset()).intValue());
             query.setMaxResults(page.getPageSize());
         }
         Page<Perusteprojekti> result = new PageImpl<>(
